requirements.txt
coverage==6.5.0
Django==4.0.5
django-allauth==0.51.0
django-celery-beat==2.3.0
django-csp==3.7
django-ipware==4.0.2
django-pandas==0.6.6
django-recaptcha==3.0.0
django-richtextfield==1.6.1
django-sendfile2==0.7.0
django-sslserver==0.22
gunicorn==20.1.0
numpy==1.23.3
phonenumbers==8.12.51
psycopg2-binary==2.9.3
pyenchant==3.2.2
reportlab==3.6.11
scikit-learn==1.1.2
squareup==22.0.0.20220921